+++ src/components/settings/forms/appearance.jsx
@@ -23,6 +23,7 @@ import {
   updateActualUserPreferences,
   setNSFWVisibility,
   setBetaChannel,
+  setUIScale,
 } from '../../../redux/action-creators';
 import settingsStyles from '../settings.module.scss';
 import { PreventPageLeaving } from '../../prevent-page-leaving';
@@ -37,6 +38,7 @@ export default function AppearanceForm() {
   const userData = useSelector((state) => state.user);
   const isNSFWVisible = useSelector((state) => state.isNSFWVisible);
   const isBetaChannel = useSelector((state) => state.betaChannel);
+  const uiScale = useSelector((state) => state.uiScale);
   const formStatus = useSelector((state) => state.settingsForms.displayPrefsStatus);
 
   useEffect(() => {
@@ -55,10 +57,10 @@ export default function AppearanceForm() {
   const form = useForm(
     useMemo(
       () => ({
-        initialValues: initialValues({ ...userData, isNSFWVisible, isBetaChannel }),
+        initialValues: initialValues({ ...userData, isNSFWVisible, isBetaChannel, uiScale }),
         onSubmit: onSubmit(dispatch),
       }),
-      [dispatch, isNSFWVisible, userData, isBetaChannel],
+      [dispatch, isNSFWVisible, userData, isBetaChannel, uiScale],
     ),
   );
 
@@ -76,6 +78,7 @@ export default function AppearanceForm() {
   const timeAmPm = useField('timeAmPm', form.form);
   const timeAbsolute = useField('timeAbsolute', form.form);
   const enableBeta = useField('enableBeta', form.form);
+  const uiScaleField = useField('uiScale', form.form);
 
   return (
     <form onSubmit={form.handleSubmit}>
@@ -201,6 +204,39 @@ export default function AppearanceForm() {
               comments
             </label>
           </div>
+        </div>
+      </section>
+
+      <section className={settingsStyles.formSection}>
+        <h4 id="scale">Text scale</h4>
+
+        <div className="form-group">
+          <p>
+            Adjust the scale of text (<strong>{uiScaleField.input.value}%</strong>):
+          </p>
+          <p>
+            <input
+              className={styles.scaleRangeInput}
+              type="range"
+              min="80"
+              max="200"
+              step="5"
+              {...uiScaleField.input}
+            />
+          </p>
+          <p>Sample text:</p>
+          <p
+            className={styles.scaleSample}
+            style={{ fontSize: `${(14 * uiScaleField.input.value) / 100}px` }}
+          >
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
+            incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
+            exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
+          </p>
+          <p className="help-block">
+            <Icon icon={faExclamationTriangle} /> This setting is saved locally in your web browser.
+            It can be different for each browser and each device that you use.
+          </p>
         </div>
       </section>
 
@@ -345,6 +381,7 @@ function initialValues({
   preferences: backend,
   isNSFWVisible,
   isBetaChannel,
+  uiScale,
 }) {
   return {
     useYou: frontend.displayNames.useYou,
@@ -361,6 +398,7 @@ function initialValues({
     timeAmPm: frontend.timeDisplay.amPm ? '1' : '0',
     timeAbsolute: frontend.timeDisplay.absolute ? '1' : '0',
     enableBeta: isBetaChannel,
+    uiScale,
   };
 }
 
@@ -371,6 +409,7 @@ function onSubmit(dispatch) {
       (dispatch) => {
         dispatch(setNSFWVisibility(!values.hideNSFWContent));
         dispatch(setBetaChannel(values.enableBeta));
+        dispatch(setUIScale(values.uiScale));
       },
     );
 }