+++ CHANGELOG.md
@@ -6,12 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.101.0] - Not released
+### Added
+- User can adjust the site font size in settings' Appearance tab. This setting
+  is saved locally in web browser and can be different for each browser and each
+  device. The available font size range is 80% - 150% relative to the default
+  size.
+- Notification messages for the 'backlink_in_post' and 'backlink_in_comment'
+  events.
+
 ### Changed
 - Post can now have an empty body if it contains one or more attachments.
 
-### Added
-- Notification messages for the 'backlink_in_post' and 'backlink_in_comment'
-  events.
+### Fixed
+- Fix the mobile Chrome unwanted font boosting
 
 ## [1.100.0] - 2021-07-30
 ### Added